feat: completed on M1

+++ srcs/requirements/wordpress/tools/init-wp.sh
@@ -1,44 +1,47 @@
 #!/bin/bash
 
-# # WordPress 설치 여부 확인
-# if [ ! -f /var/www/html/index.php ]; then
-#     echo "WordPress가 설치되어 있지 않습니다. 설치를 시작합니다."
+set -e
 
-#     # 1. /tmp 디렉토리로 이동
-#     cd /tmp
+# `wp-config.php`가 없으면 설치 진행
+if [ ! -f "/var/www/html/wp-config.php" ]; then
+    echo "wp-config.php being created..."
 
-#     # 2. WordPress 최신 버전 다운로드
-#     wget https://wordpress.org/latest.tar.gz
-
-#     # 3. 다운로드한 tar 파일을 추출
-#     tar -xvf latest.tar.gz
-
-#     # 4. 추출한 WordPress 파일을 /var/www/html/로 복사
-#     cp -r wordpress/* /var/www/html/
-
-#     cp /tmp/pool.d/www.conf /etc/php/7.4/fpm/pool.d/www.conf
-#     # mv /tmp/wp-config.php /var/www/html/wp-config.php
+    # 데이터베이스 연결 설정
+    cp -p /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
+    sed -i "s/username_here/$WORDPRESS_DB_USER/g" /var/www/html/wp-config.php
+    sed -i "s/password_here/$WORDPRESS_DB_PASSWORD/g" /var/www/html/wp-config.php
+    sed -i "s/localhost/$WORDPRESS_DB_HOST/g" /var/www/html/wp-config.php
+    sed -i "s/database_name_here/$WORDPRESS_DB_NAME/g" /var/www/html/wp-config.php
     
-#     # 5. /var/www/html/의 소유자와 그룹을 www-data로 변경
-#     chown -R www-data:www-data /var/www/html/
+    echo "wp-config.php is created successfully."
+else
+    echo "wp-config.php is already created." 
+fi
 
-#     # 6. /var/www/html/의 권한을 755로 설정
-#     chmod -R 755 /var/www/html/
+# wordpress 가 없으면 설치 진행
+if ! wp core is-installed --path="/var/www/html" --allow-root; then
+    echo "WordPress not installed. Setting up..."
 
-#     echo "WordPress 설치가 완료되었습니다."
-# else
-#     echo "WordPress가 이미 설치되어 있습니다. 설치를 건너뜁니다."
-# fi
+    #wordpress 설치 
+    wp core install \
+        --url="https://localhost" \
+        --title=${WORDPRESS_SITE_TITLE} \
+        --admin_user=${WORDPRESS_ADMIN_USER} \
+        --admin_password=${WORDPRESS_ADMIN_PASSWORD} \
+        --admin_email=${WORDPRESS_ADMIN_EMAIL} \
+        --path="/var/www/html" \
+        --allow-root
+    #wordpress user 생성
+    wp user create $WORDPRESS_USER $WORDPRESS_USER_EMAIL \
+					--user_pass=$WORDPRESS_PASSWORD \
+                    --path="/var/www/html" \
+					--allow-root
 
-cp -p /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
-sed -i "s/username_here/$WORDPRESS_DB_USER/g" /var/www/html/wp-config.php
-sed -i "s/password_here/$WORDPRESS_DB_PASSWORD/g" /var/www/html/wp-config.php
-sed -i "s/localhost/$WORDPRESS_DB_HOST/g" /var/www/html/wp-config.php
-sed -i "s/database_name_here/$WORDPRESS_DB_NAME/g" /var/www/html/wp-config.php
+    echo "WordPress installed successfully."
+else
+    echo "WordPress already installed." 
+fi
 
 
-# # 7. PHP 런타임에 필요한 디렉토리 생성
-# mkdir -p /run/php
-
-# 전달받은 명령어 실행
+# exec CMD
 exec "$@"
